Added files via upload
This is the Model class that I am currently using to test the Ai function. The copy method is not complete or practical and neither is the playerWinSimple() method. The playerWinSimple() method is used to test a 1 X 2 board. This is helping in testing the logic in the Ai function.